feat(mobile): 重做播放器无级手势并精简主题设置
原有移动端亮度与音量手势以 5% 台阶量化，画面反馈位于侧边，且高频音量更新会逐帧触发播放器 React 状态变更。移动端设置页同时重复暴露了显式亮暗模式入口。

本次将亮度与音量改为按完整画面高度映射的连续调节，并提供中央实时 HUD；音量拖动直接更新媒体元素，Android 亮度以浮点值贯通前端、Rust 与 Kotlin，快速原生写入在 IPC 前合并。移动端设置页隐藏显式主题入口，桌面端主题切换保持不变。

- 支持 Android、iOS 的左右分区纵向手势与会话级亮度清理
- HUD 使用作用域 GSAP 动画并遵循 prefers-reduced-motion
- 版本升级至 0.16.0，补充连续值、原生桥和设置入口回归测试
- 验证通过：bun run check、321 项前端测试、bun run build、cargo test（295 项通过，6 项网络 smoke 忽略）
- Android Kotlin 未执行真机构建，需后续设备回归

+++ src-tauri/src/commands/android_player_controls.rs
@@ -57,7 +57,7 @@ mod android {
     #[tauri::command]
     pub async fn android_player_controls_set_brightness(
         controls: State<'_, AndroidPlayerControls>,
-        value: i32,
+        value: f64,
     ) -> AppResult<Value> {
         run(controls, "setBrightness", json!({ "value": value })).await
     }
@@ -111,7 +111,7 @@ mod fallback {
     }
 
     #[tauri::command]
-    pub async fn android_player_controls_set_brightness(_value: i32) -> AppResult<Value> {
+    pub async fn android_player_controls_set_brightness(_value: f64) -> AppResult<Value> {
         Err(unsupported())
     }
